+++ workers/tests/unittests/test_validators.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 import pytest
 
 from ohsome_quality_analyst.utils.exceptions import (
@@ -5,8 +7,10 @@ from ohsome_quality_analyst.utils.exceptions import (
     GeoJSONGeometryTypeError,
     GeoJSONObjectTypeError,
     IndicatorTopicCombinationError,
+    SizeRestrictionError,
 )
 from ohsome_quality_analyst.utils.validators import (
+    validate_area,
     validate_geojson,
     validate_indicator_topic_combination,
 )
@@ -62,3 +66,23 @@ def test_validate_indicator_topic_combination():
 def test_validate_indicator_topic_combination_invalid():
     with pytest.raises(IndicatorTopicCombinationError):
         validate_indicator_topic_combination("minimal", "building-count")
+
+
+@mock.patch.dict(
+    "os.environ",
+    {"OQT_GEOM_SIZE_LIMIT": "1000"},
+    clear=True,
+)
+def test_validate_area(feature_germany_heidelberg):
+    # expect not exceptions
+    validate_area(feature_germany_heidelberg)
+
+
+@mock.patch.dict(
+    "os.environ",
+    {"OQT_GEOM_SIZE_LIMIT": "1"},
+    clear=True,
+)
+def test_validate_area_exception(feature_germany_heidelberg):
+    with pytest.raises(SizeRestrictionError):
+        validate_area(feature_germany_heidelberg)